requirements.txt
asgiref==3.11.0
channels==4.1.0
channels-redis==4.2.0
cffi==2.0.0
cryptography==46.0.3
daphne==4.1.2
Django==4.2.26
django-appconf==1.2.0
django-cors-headers==4.9.0
django-cryptography==1.1
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
drf-writable-nested==0.7.2
pillow==12.0.0
pycparser==2.23
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
